@@ -3,10 +3,10 @@
     "version": "1.0.0",
     "dependencies": {
         "rxjs": "~6.6.3",
-        "@angular/core": "21.0.2",
-        "@angular/forms": "21.0.2",
+        "@angular/core": "21.1.2",
+        "@angular/forms": "21.1.2",
         "vite": "7.2.7",
-        "primeng": "^21.0.2",
+        "primeng": "^21.1.1",
         "@nx/vite": "21.6.9"
     },
     "type": "module",
