@@ -46,17 +46,17 @@
     },
     "private": false,
     "dependencies": {
-        "@angular-devkit/core": "21.0.2",
-        "@angular/animations": "21.0.2",
-        "@angular/cdk": "21.0.2",
-        "@angular/common": "21.0.2",
-        "@angular/compiler": "21.0.2",
-        "@angular/core": "21.0.2",
-        "@angular/elements": "21.0.2",
-        "@angular/forms": "21.0.2",
-        "@angular/platform-browser": "21.0.2",
-        "@angular/platform-browser-dynamic": "21.0.2",
-        "@angular/router": "21.0.2",
+        "@angular-devkit/core": "21.1.2",
+        "@angular/animations": "21.1.2",
+        "@angular/cdk": "21.1.2",
+        "@angular/common": "21.1.2",
+        "@angular/compiler": "21.1.2",
+        "@angular/core": "21.1.2",
+        "@angular/elements": "21.1.2",
+        "@angular/forms": "21.1.2",
+        "@angular/platform-browser": "21.1.2",
+        "@angular/platform-browser-dynamic": "21.1.2",
+        "@angular/router": "21.1.2",
         "@ctrl/tinycolor": "^3.1.7",
         "@date-fns/tz": "^1.4.0",
         "@jitsu/sdk-js": "^3.1.5",
@@ -125,7 +125,7 @@
         "node-fetch": "^2.6.1",
         "primeflex": "3.3.1",
         "primeicons": "7.0.0",
-        "primeng": "^21.0.4",
+        "primeng": "^21.1.1",
         "react": "18.3.1",
         "react-dom": "18.3.1",
         "regenerator-runtime": "^0.13.9",
@@ -157,16 +157,16 @@
         "@analytics/queue-utils": "^0.1.3",
         "@analytics/router-utils": "^0.1.1",
         "@analytics/url-utils": "^0.2.3",
-        "@angular-devkit/architect": "0.2100.2",
-        "@angular-devkit/build-angular": "21.0.2",
-        "@angular-devkit/schematics": "21.0.2",
+        "@angular-devkit/architect": "0.2101.2",
+        "@angular-devkit/build-angular": "21.1.2",
+        "@angular-devkit/schematics": "21.1.2",
         "@angular-eslint/eslint-plugin": "^21.0.0",
         "@angular-eslint/eslint-plugin-template": "^21.0.0",
         "@angular-eslint/schematics": "^21.0.0",
         "@angular-eslint/template-parser": "^21.0.0",
-        "@angular/cli": "21.0.2",
-        "@angular/compiler-cli": "21.0.2",
-        "@angular/language-service": "21.0.2",
+        "@angular/cli": "21.1.2",
+        "@angular/compiler-cli": "21.1.2",
+        "@angular/language-service": "21.1.2",
         "@babel/core": "^7.14.5",
         "@babel/plugin-proposal-class-properties": "^7.18.6",
         "@babel/plugin-proposal-private-methods": "^7.18.6",
@@ -195,7 +195,7 @@
         "@nxext/stencil": "21.0.0",
         "@playwright/test": "^1.36.0",
         "@rollup/plugin-url": "^8.0.2",
-        "@schematics/angular": "21.0.2",
+        "@schematics/angular": "21.1.2",
         "@stencil/core": "^4.39.0",
         "@stencil/sass": "^3.2.3",
         "@storybook/addon-docs": "8.6.14",
