@@ -27,21 +27,21 @@ import {
 } from '@dotcms/data-access';
 import { SiteService } from '@dotcms/dotcms-js';
 import {
+    CONTAINER_SOURCE,
     DotActionBulkResult,
+    DotActionMenuItem,
     DotBulkFailItem,
     DotContainer,
     DotContentState,
-    CONTAINER_SOURCE,
     DotMessageSeverity,
-    DotMessageType,
-    DotActionMenuItem
+    DotMessageType
 } from '@dotcms/dotcms-models';
 import {
     DotActionMenuButtonComponent,
     DotAddToBundleComponent,
+    DotContentletStatusChipComponent,
     DotMessagePipe,
-    DotRelativeDatePipe,
-    DotContentletStatusChipComponent
+    DotRelativeDatePipe
 } from '@dotcms/ui';
 
 import { DotContainerListResolver } from './dot-container-list-resolver.service';
@@ -56,7 +56,6 @@ import { DotPortletBaseComponent } from '../../../view/components/dot-portlet-ba
 @Component({
     selector: 'dot-container-list',
     templateUrl: './container-list.component.html',
-    styleUrls: ['./container-list.component.css'],
     imports: [
         CommonModule,
         DotPortletBaseComponent,
@@ -197,6 +196,10 @@ export class ContainerListComponent implements OnDestroy {
     handleActionMenuOpen(event: MouseEvent): void {
         this.updateSelectedContainers();
         this.actionsMenu.toggle(event);
+    }
+
+    handleSelectionChange(): void {
+        this.updateSelectedContainers();
     }
 
     /**
