@@ -1,10 +1,8 @@
 import { enableProdMode } from '@angular/core';
 import { bootstrapApplication } from '@angular/platform-browser';
 
-// import { defineCustomElements } from '@dotcms/dotcms-webcomponents/loader';
-//
-// import { defineCustomElements } from '@dotcms/dotcms-webcomponents/loader';
-//
+import { defineCustomElements } from '@dotcms/dotcms-webcomponents/loader';
+
 import { AppComponent } from './app/app.component';
 import { appConfig } from './app/app.config';
 import { environment } from './environments/environment';
@@ -14,4 +12,4 @@ if (environment.production) {
 }
 
 bootstrapApplication(AppComponent, appConfig);
-//defineCustomElements();
+defineCustomElements();
