@@ -904,7 +904,7 @@ describe('DotFormComponent', () => {
         describe('Historical Version UI Elements', () => {
             it('should hide lock controls when viewing historical version', () => {
                 // Initially lock controls should be visible
-                const lockControls = spectator.query('.dot-edit-content-actions__lock');
+                const lockControls = spectator.query(byTestId('content-lock-controls'));
                 expect(lockControls).toBeTruthy();
 
                 // Simulate loading a historical version using the store's public method
@@ -912,7 +912,7 @@ describe('DotFormComponent', () => {
                 spectator.detectChanges();
 
                 // Lock controls should be hidden
-                const lockControlsAfter = spectator.query('.dot-edit-content-actions__lock');
+                const lockControlsAfter = spectator.query(byTestId('content-lock-controls'));
                 expect(lockControlsAfter).toBeFalsy();
             });
 
@@ -993,7 +993,7 @@ describe('DotFormComponent', () => {
                 expect(store.isViewingHistoricalVersion()).toBe(false);
                 //TODO: enable this when all fields have disable state expect(component.form.enabled).toBe(true);
 
-                const lockControls = spectator.query('.dot-edit-content-actions__lock');
+                const lockControls = spectator.query(byTestId('content-lock-controls'));
                 const workflowActions = spectator.query(byTestId('workflow-actions'));
                 const restoreButton = spectator.query(
                     byTestId('restore-historical-version-button')
@@ -1010,7 +1010,7 @@ describe('DotFormComponent', () => {
                 // Check historical view state
                 //TODO: enable this when all fields have disable state expect(component.form.disabled).toBe(true);
 
-                const lockControlsAfter = spectator.query('.dot-edit-content-actions__lock');
+                const lockControlsAfter = spectator.query(byTestId('content-lock-controls'));
                 const workflowActionsAfter = spectator.query(byTestId('workflow-actions'));
                 const restoreButtonAfter = spectator.query(
                     byTestId('restore-historical-version-button')
@@ -1036,7 +1036,7 @@ describe('DotFormComponent', () => {
                 // Check normal view state
                 expect(component.form.enabled).toBe(true);
 
-                const lockControls = spectator.query('.dot-edit-content-actions__lock');
+                const lockControls = spectator.query(byTestId('content-lock-controls'));
                 const workflowActions = spectator.query(byTestId('workflow-actions'));
                 const restoreButton = spectator.query(
                     byTestId('restore-historical-version-button')
