@@ -11,8 +11,8 @@ import {
 
 import { AvatarModule } from 'primeng/avatar';
 import { ButtonModule } from 'primeng/button';
-import { ChipModule } from 'primeng/chip';
 import { MenuModule } from 'primeng/menu';
+import { TagModule } from 'primeng/tag';
 import { TooltipModule } from 'primeng/tooltip';
 
 import { DotMessageService } from '@dotcms/data-access';
@@ -42,8 +42,8 @@ import {
         CommonModule,
         AvatarModule,
         ButtonModule,
-        ChipModule,
         MenuModule,
+        TagModule,
         TooltipModule,
         DotGravatarDirective,
         DotMessagePipe,
