@@ -21,9 +21,9 @@ import { TreeModule } from 'primeng/tree';
 import { DotMessageService } from '@dotcms/data-access';
 import { ComponentStatus } from '@dotcms/dotcms-models';
 import {
+    DotCollapseBreadcrumbComponent,
     DotEmptyContainerComponent,
-    PrincipalConfiguration,
-    DotCollapseBreadcrumbComponent
+    PrincipalConfiguration
 } from '@dotcms/ui';
 
 import { CATEGORY_FIELD_EMPTY_MESSAGES } from '../../../../models/dot-edit-content-field.constant';
@@ -55,10 +55,9 @@ const MINIMUM_CATEGORY_WITHOUT_SCROLLING = 3;
         DotEmptyContainerComponent
     ],
     templateUrl: './dot-category-field-category-list.component.html',
-    styleUrl: './dot-category-field-category-list.component.scss',
     changeDetection: ChangeDetectionStrategy.OnPush,
     host: {
-        class: 'category-list__wrapper'
+        class: 'overflow-auto h-full flex flex-col rounded-md'
     }
 })
 export class DotCategoryFieldCategoryListComponent {
