@@ -1,10 +1,10 @@
 import { InferInputSignals } from '@ngneat/spectator';
 import { createComponentFactory, mockProvider, Spectator } from '@ngneat/spectator/jest';
-import { throwError } from 'rxjs';
+import { of, throwError } from 'rxjs';
 
 import { HttpClient } from '@angular/common/http';
 import { computed, signal } from '@angular/core';
-import { fakeAsync, tick } from '@angular/core/testing';
+import { fakeAsync, flushMicrotasks, tick } from '@angular/core/testing';
 import { FormGroup } from '@angular/forms';
 
 import { Accordion, AccordionModule } from 'primeng/accordion';
@@ -148,7 +148,7 @@ describe('DotUveStyleEditorFormComponent', () => {
             activeContentlet: signal(null),
             graphqlResponse: graphqlResponseSignal,
             $customGraphqlResponse: customGraphqlResponseComputed,
-            saveStyleEditor: jest.fn(),
+            saveStyleEditor: jest.fn().mockReturnValue(of({})),
             rollbackGraphqlResponse: jest.fn().mockReturnValue(true),
             addHistory: jest.fn(),
             setGraphqlResponse: jest.fn((response: DotCMSPageAsset | null) => {
@@ -172,10 +172,14 @@ describe('DotUveStyleEditorFormComponent', () => {
             expect(spectator.component).toBeTruthy();
         });
 
-        it('should initialize form when schema is provided', () => {
+        it('should initialize form when schema is provided', fakeAsync(() => {
+            // Flush microtasks to allow form to be built
+            flushMicrotasks();
+            spectator.detectChanges();
+
             expect(spectator.component.$form()).toBeTruthy();
             expect(spectator.component.$form()).toBeInstanceOf(FormGroup);
-        });
+        }));
 
         it('should compute sections from schema', () => {
             const sections = spectator.component.$sections();
@@ -186,6 +190,12 @@ describe('DotUveStyleEditorFormComponent', () => {
     });
 
     describe('form structure', () => {
+        beforeEach(fakeAsync(() => {
+            // Ensure form is built before testing structure
+            flushMicrotasks();
+            spectator.detectChanges();
+        }));
+
         it('should render accordion with sections', () => {
             const accordion = spectator.query(Accordion);
             expect(accordion).toBeDefined();
@@ -198,6 +208,12 @@ describe('DotUveStyleEditorFormComponent', () => {
     });
 
     describe('field rendering', () => {
+        beforeEach(fakeAsync(() => {
+            // Ensure form is built before testing field rendering
+            flushMicrotasks();
+            spectator.detectChanges();
+        }));
+
         it('should render input field component', () => {
             const inputField = spectator.query('dot-uve-style-editor-field-input');
             expect(inputField).toBeTruthy();
@@ -220,7 +236,7 @@ describe('DotUveStyleEditorFormComponent', () => {
     });
 
     describe('initial values from contentlet styleProperties', () => {
-        it('should use styleProperties from activeContentlet when available', () => {
+        it('should use styleProperties from activeContentlet when available', fakeAsync(() => {
             const styleProperties = {
                 'font-size': 20,
                 'font-family': 'Helvetica',
@@ -260,6 +276,10 @@ describe('DotUveStyleEditorFormComponent', () => {
             });
             spectator.detectChanges();
 
+            // Flush microtasks to allow form to be built
+            flushMicrotasks();
+            spectator.detectChanges();
+
             const form = spectator.component.$form();
             expect(form?.get('font-size')?.value).toBe(20);
             expect(form?.get('font-family')?.value).toBe('Helvetica');
@@ -268,7 +288,7 @@ describe('DotUveStyleEditorFormComponent', () => {
             const textDecorationGroup = form?.get('text-decoration') as FormGroup;
             expect(textDecorationGroup.get('underline')?.value).toBe(false);
             expect(textDecorationGroup.get('overline')?.value).toBe(true);
-        });
+        }));
     });
 
     describe('rollback and form restoration', () => {
@@ -317,6 +337,10 @@ describe('DotUveStyleEditorFormComponent', () => {
             });
             spectator.detectChanges();
 
+            // Flush microtasks to allow form to be built
+            flushMicrotasks();
+            spectator.detectChanges();
+
             let form = spectator.component.$form();
             expect(form?.get('font-size')?.value).toBe(16);
 
@@ -331,9 +355,14 @@ describe('DotUveStyleEditorFormComponent', () => {
             );
 
             // Change form value (this triggers the save flow)
+            // activeContentlet is captured at the time of form change (before debounce)
             form?.patchValue({ 'font-size': 20 });
             tick(STYLE_EDITOR_DEBOUNCE_TIME + 100); // Wait for debounce + error handling
             spectator.detectChanges(); // Ensure change detection runs after rollback
+
+            // Flush microtasks after rollback to allow form to be rebuilt
+            flushMicrotasks();
+            spectator.detectChanges();
 
             // Get the NEW form reference after rollback (form is rebuilt, not patched)
             form = spectator.component.$form();
@@ -353,6 +382,10 @@ describe('DotUveStyleEditorFormComponent', () => {
             });
             spectator.detectChanges();
 
+            // Flush microtasks to allow form to be built
+            flushMicrotasks();
+            spectator.detectChanges();
+
             const rolledBackResponse = createMockGraphQLResponse(16);
 
             // Mock saveStyleEditor to always fail and rollback to 16
@@ -364,10 +397,15 @@ describe('DotUveStyleEditorFormComponent', () => {
             );
 
             // First failure: change from 16 to 20, then fail
+            // activeContentlet is captured at the time of form change (before debounce)
             let form = spectator.component.$form();
             form?.patchValue({ 'font-size': 20 });
             tick(STYLE_EDITOR_DEBOUNCE_TIME + 100);
             spectator.detectChanges(); // Ensure change detection runs after rollback
+
+            // Flush microtasks after first rollback to allow form to be rebuilt
+            flushMicrotasks();
+            spectator.detectChanges();
 
             // Get the NEW form reference after first rollback (form is rebuilt)
             form = spectator.component.$form();
@@ -375,10 +413,15 @@ describe('DotUveStyleEditorFormComponent', () => {
 
             // Second failure: Get fresh form reference before patching
             // This ensures we're patching the current form instance
+            // activeContentlet is captured again at the time of form change
             form = spectator.component.$form();
             form?.patchValue({ 'font-size': 24 });
             tick(STYLE_EDITOR_DEBOUNCE_TIME + 100);
             spectator.detectChanges(); // Ensure change detection runs after rollback
+
+            // Flush microtasks after second rollback to allow form to be rebuilt
+            flushMicrotasks();
+            spectator.detectChanges();
 
             // Get the NEW form reference after second rollback
             form = spectator.component.$form();
@@ -395,6 +438,10 @@ describe('DotUveStyleEditorFormComponent', () => {
             });
             spectator.detectChanges();
 
+            // Flush microtasks to allow form to be built
+            flushMicrotasks();
+            spectator.detectChanges();
+
             // Get initial form reference
             const initialForm = spectator.component.$form();
             expect(initialForm?.get('font-size')?.value).toBe(16);
@@ -409,9 +456,14 @@ describe('DotUveStyleEditorFormComponent', () => {
             );
 
             // Change form value to trigger save and rollback
+            // activeContentlet is captured at the time of form change (before debounce)
             initialForm?.patchValue({ 'font-size': 20 });
             tick(STYLE_EDITOR_DEBOUNCE_TIME + 100);
             spectator.detectChanges(); // Ensure change detection runs after rollback
+
+            // Flush microtasks after rollback to allow form to be rebuilt
+            flushMicrotasks();
+            spectator.detectChanges();
 
             // Get form reference after rollback
             const rebuiltForm = spectator.component.$form();
@@ -419,6 +471,70 @@ describe('DotUveStyleEditorFormComponent', () => {
             // Verify form was REBUILT (new instance), not patched
             expect(rebuiltForm).not.toBe(initialForm);
             expect(rebuiltForm?.get('font-size')?.value).toBe(16);
+        }));
+
+        it('should capture activeContentlet at form change time and pass it to saveStyleProperties', fakeAsync(() => {
+            // Create component with activeContentlet
+            spectator = createComponent({
+                props: {
+                    ['schema' as keyof InferInputSignals<DotUveStyleEditorFormComponent>]:
+                        createMockSchema()
+                }
+            });
+            spectator.detectChanges();
+
+            // Flush microtasks to allow form to be built
+            flushMicrotasks();
+            spectator.detectChanges();
+
+            const initialActiveContentlet = mockUveStore.activeContentlet();
+            expect(initialActiveContentlet).toBeTruthy();
+
+            // Mock saveStyleEditor to succeed and verify it receives the correct activeContentlet
+            mockUveStore.saveStyleEditor.mockReturnValue(of({}));
+
+            // Change form value (this triggers the save flow)
+            // activeContentlet is captured at the time of form change (before debounce)
+            const form = spectator.component.$form();
+            form?.patchValue({ 'font-size': 20 });
+
+            // Change activeContentlet AFTER form change but BEFORE debounce completes
+            // This tests that the original activeContentlet is captured and used
+            const newActiveContentlet = {
+                contentlet: {
+                    identifier: 'new-test-id',
+                    inode: 'new-test-inode',
+                    title: 'New Test',
+                    contentType: 'test-content-type',
+                    dotStyleProperties: { 'font-size': 30 }
+                },
+                container: {
+                    acceptTypes: 'test',
+                    identifier: 'new-test-container',
+                    maxContentlets: 1,
+                    uuid: 'new-test-uuid'
+                },
+                language_id: '1',
+                pageContainers: [],
+                pageId: 'new-test-page'
+            };
+            mockUveStore.activeContentlet.set(newActiveContentlet);
+
+            // Wait for debounce to complete
+            tick(STYLE_EDITOR_DEBOUNCE_TIME + 100);
+            spectator.detectChanges();
+
+            // Verify saveStyleEditor was called with the ORIGINAL activeContentlet
+            // (captured at form change time, not the new one)
+            expect(mockUveStore.saveStyleEditor).toHaveBeenCalledTimes(1);
+            const saveCall = mockUveStore.saveStyleEditor.mock.calls[0][0];
+            expect(saveCall.contentletIdentifier).toBe(
+                initialActiveContentlet?.contentlet.identifier
+            );
+            expect(saveCall.containerIdentifier).toBe(
+                initialActiveContentlet?.container.identifier
+            );
+            expect(saveCall.pageId).toBe(initialActiveContentlet?.pageId);
         }));
     });
 });
