@@ -23,7 +23,10 @@ import { DotMessagePipe } from '../../dot-message/dot-message.pipe';
     selector: 'dot-form-dialog',
     imports: [ButtonModule, FocusTrapModule, DotMessagePipe],
     templateUrl: './dot-form-dialog.component.html',
-    styleUrls: ['./dot-form-dialog.component.scss']
+    host: {
+        class: 'flex min-h-0 flex-1 flex-col',
+        style: 'height: 100%'
+    }
 })
 export class DotFormDialogComponent implements OnInit, OnDestroy {
     private dynamicDialog = inject(DynamicDialogRef);
